poc__llm_search_attributes_vector/api/requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.5; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
altair==5.3.0; python_version >= '3.8'
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.3.0; python_version >= '3.8'
asn1crypto==1.5.1
attrs==23.2.0; python_version >= '3.7'
blinker==1.7.0; python_version >= '3.8'
boto3==1.34.91; python_version >= '3.8'
botocore==1.34.91; python_version >= '3.8'
cachetools==5.3.3; python_version >= '3.7'
certifi==2024.2.2; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cryptography==42.0.5; python_version >= '3.7'
dataclasses-json==0.6.4; python_version >= '3.7' and python_version < '4.0'
decorator==5.1.1; python_version >= '3.5'
deprecation==2.1.0
distro==1.9.0; python_version >= '3.6'
fastapi==0.110.2; python_version >= '3.8'
filelock==3.13.4; python_version >= '3.8'
frozenlist==1.4.1; python_version >= '3.8'
gitdb==4.0.11; python_version >= '3.7'
gitpython==3.1.43; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httpx==0.27.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
jinja2==3.1.3; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==2.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonschema==4.21.1; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
lancedb==0.6.10; python_version >= '3.8'
langchain==0.1.16; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-community==0.0.34; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-core==0.1.45; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-openai==0.1.3; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-text-splitters==0.0.1; python_version < '4.0' and python_full_version >= '3.8.1'
langsmith==0.1.50; python_version < '4.0' and python_full_version >= '3.8.1'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
marshmallow==3.21.1; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
multidict==6.0.5; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
numpy==1.26.4; python_version >= '3.9'
openai==1.23.4; python_full_version >= '3.7.1'
orjson==3.10.1; python_version >= '3.8'
overrides==7.7.0; python_version >= '3.6'
packaging==23.2; python_version >= '3.7'
pandas==2.2.2; python_version >= '3.9'
pillow==10.3.0; python_version >= '3.8'
platformdirs==4.2.1; python_version >= '3.8'
polars==0.20.22; python_version >= '3.8'
protobuf==4.25.3; python_version >= '3.8'
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyarrow==15.0.0; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pydantic==2.7.1; python_version >= '3.8'
pydantic-core==2.18.2; python_version >= '3.8'
pydeck==0.9.0b0; python_version >= '3.8'
pygments==2.17.2; python_version >= '3.7'
pyjwt==2.8.0; python_version >= '3.7'
pylance==0.10.12; python_version >= '3.8'
pyopenssl==24.1.0; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
ratelimiter==1.2.0.post0
referencing==0.35.0; python_version >= '3.8'
regex==2024.4.16; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
retry==0.9.2
rich==13.7.1; python_full_version >= '3.7.0'
rpds-py==0.18.0; python_version >= '3.8'
s3transfer==0.10.1; python_version >= '3.8'
semver==3.0.2; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.1; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
snowflake-connector-python==3.9.1; python_version >= '3.8'
sortedcontainers==2.4.0
sqlalchemy==2.0.29; python_version >= '3.7'
starlette==0.37.2; python_version >= '3.8'
streamlit==1.33.0; python_version >= '3.8' and python_full_version != '3.9.7'
tenacity==8.2.3; python_version >= '3.7'
tiktoken==0.6.0; python_version >= '3.8'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomlkit==0.12.4; python_version >= '3.7'
toolz==0.12.1; python_version >= '3.7'
tornado==6.4; python_version >= '3.8'
tqdm==4.66.2; python_version >= '3.7'
typing-extensions==4.11.0; python_version >= '3.8'
typing-inspect==0.9.0
tzdata==2024.1; python_version >= '2'
urllib3==2.2.1; python_version >= '3.8'
uvicorn==0.29.0; python_version >= '3.8'
yarl==1.9.4; python_version >= '3.7'
